% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{get_parameter_order}
\alias{get_parameter_order}
\title{Internal variance parameter order}
\usage{
get_parameter_order(prior_obj)
}
\arguments{
\item{prior_obj}{Object of class \code{mmp_prior}, see \link[makemyprior]{make_prior}.}
}
\value{
Names of the random effects in the model in the order the prior object reads them.
}
\description{
Returns the internal order of the variance parameters related to each random effect in the model.
}
\examples{
ex_model <- makemyprior_example_model()
get_parameter_order(ex_model)

}
