% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(x)
}
\arguments{
\item{x}{Value between 0 and 1, or vector of such values}
}
\value{
logit value
}
\description{
Calculates logit, log(x/(1-x)) = log(x) - log(1-x)
}
\examples{
logit(0.5)
logit(seq(0, 1, 0.1))

}
