% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{makemyprior_models}
\alias{makemyprior_models}
\title{List available priors, latent models and likelihoods}
\usage{
makemyprior_models(type = c("prior", "latent", "likelihood"), select = NULL)
}
\arguments{
\item{type}{Which of priors, latent models and likelihoods to list. Options are
\code{prior}, \code{latent} and \code{likelihood}}

\item{select}{Which in each group to show details about.
\code{NULL} gives only list what exists (default),
\code{all} gives detailed information about everything in that category.
Can also ask for one or more specific priors/latent models/likelihoods}
}
\value{
None.
}
\description{
Prints available priors, latent models and likelihoods to use with \link[makemyprior]{make_prior}.
}
\examples{
makemyprior_models("prior", c("pc0", "pc1"))
makemyprior_models("latent")
makemyprior_models("likelihood", "all")

}
