% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{makemyprior_plotting}
\alias{makemyprior_plotting}
\title{List of available plotting functions}
\usage{
makemyprior_plotting()
}
\value{
None.
}
\description{
Directs to this help page.
}
\details{
The available plotting functions are:
\itemize{
  \item{\link[makemyprior]{plot_marginal_prior}}
  \item{\link[makemyprior]{plot_posterior_fixed}}
  \item{\link[makemyprior]{plot_posterior_precision}}
  \item{\link[makemyprior]{plot_posterior_stan}}
  \item{\link[makemyprior]{plot_posterior_stdev}}
  \item{\link[makemyprior]{plot_posterior_variance}}
  \item{\link[makemyprior]{plot_prior}}
  \item{\link[makemyprior]{plot_several_posterior_stan}}
  \item{\link[makemyprior]{plot_tree_structure}}
}
In addition the following functions can be used to extract posterior samples of
effects and parameters:
\itemize{
  \item{\link[makemyprior]{extract_posterior_effect}}
  \item{\link[makemyprior]{extract_posterior_parameter}}
}

\link[makemyprior]{eval_pc_prior} can be used to evaluate a PC prior for a weight parameter,
and \link[makemyprior]{eval_joint_prior} to evaluate the whole joint prior.
}
