% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{plot.mmp_prior}
\alias{plot.mmp_prior}
\alias{plot.mmp_inla}
\alias{plot.mmp_stan}
\title{Plotting}
\usage{
\method{plot}{mmp_prior}(x, ...)

\method{plot}{mmp_inla}(x, ...)

\method{plot}{mmp_stan}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mmp_prior}, \code{mmp_inla} or \code{mmp_stan}.}

\item{...}{Additional arguments to plotting functions. Varies with what object is sent to function.}
}
\value{
None.
}
\description{
Plotting
}
\details{
See \link[makemyprior]{plot_prior} (objects of class \code{mmp_prior}),
\link[makemyprior]{plot_posterior_stan} (objects of class \code{mmp_stan}), and
\link[makemyprior]{plot_posterior_variance} (objects of class \code{mmp_inla}),
}
\examples{
pri <- makemyprior_example_model()
plot(pri)

if (interactive() && requireNamespace("rstan")){
  res_stan <- inference_stan(ex_prior, iter = 100)
  # Note: For reliable results, increase the number of iterations (e.g., 'iter = 2000')
  plot(res_stan)
}

if (interactive() && requireNamespace("INLA")){
  res_inla <- inference_inla(pri)
 plot(res_inla)
}

}
