% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_posterior_diagnostics.R
\name{plot_prior}
\alias{plot_prior}
\title{Plotting prior distributions}
\usage{
plot_prior(obj)
}
\arguments{
\item{obj}{An object from \code{make_prior}, \code{inference_stan}, \code{inference_inla}, or \code{makemyprior_gui}}
}
\value{
A \link[ggplot2]{ggplot} with the prior distributions.
See also \link[makemyprior]{makemyprior_plotting}.
}
\description{
Function for plotting the prior distributions of the random effects on the scale of the parameters chosen
}
\examples{
ex_prior <- makemyprior_example_model()
plot_prior(ex_prior)

}
\keyword{plot}
