% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_posterior_diagnostics.R
\name{plot_tree_structure}
\alias{plot_tree_structure}
\title{Plotting the prior tree structure graph}
\usage{
plot_tree_structure(obj, nodenames)
}
\arguments{
\item{obj}{An object from \code{make_prior}, \code{inference_stan}, \code{inference_inla}, or \code{makemyprior_gui}}

\item{nodenames}{Custom names for each node (optional). Given as a named list with the default names as list names, and the
new names as list elements.
Do not need to provide all.}
}
\value{
A \link[visNetwork]{visNetwork} with the tree graph
See also \link[makemyprior]{makemyprior_plotting}.
}
\description{
Can only be used for visualization in R.
}
\examples{
ex_prior <- makemyprior_example_model()
plot_tree_structure(ex_prior)

}
\keyword{plot}
