% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{scale_precmat}
\alias{scale_precmat}
\title{Scaling precision matrix}
\usage{
scale_precmat(Q)
}
\arguments{
\item{Q}{Precision matrix}
}
\value{
Precision matrix which is now scaled to have typical variance 1.
}
\description{
Scaling a precision matrix so the corresponding covariance matrix has typical variance (geometric mean) equal to 1.
}
\examples{
scale_precmat(diag(10))

}
