% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{summary.mmp_prior}
\alias{summary.mmp_prior}
\alias{summary.mmp_inla}
\alias{summary.mmp_stan}
\title{Short summary}
\usage{
\method{summary}{mmp_prior}(object, ...)

\method{summary}{mmp_inla}(object, ...)

\method{summary}{mmp_stan}(object, ...)
}
\arguments{
\item{object}{Object of class \code{mmp_prior}, \code{mmp_inla} or \code{mmp_stan}.}

\item{...}{For \code{mmp_stan} and \code{mmp_inla}, see \link[base]{print.data.frame}.}
}
\value{
Returns summary invisible.
}
\description{
Short summary
}
\examples{
\dontrun{
pri <- makemyprior_example_model()
summary(pri)
res_stan <- inference_stan(pri)
summary(res_stan)
res_inla <- inference_inla(pri)
summary(res_inla)
}

}
