% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{typical_variance}
\alias{typical_variance}
\title{Compute the typical variance}
\usage{
typical_variance(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\value{
Typical variance.
}
\description{
Computing the typical variance (geometric mean) of a matrix.
}
\examples{
typical_variance(diag(10))

}
