% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{expit}
\alias{expit}
\title{expit}
\usage{
expit(x)
}
\arguments{
\item{x}{Real number}
}
\value{
expit value
}
\description{
Calculates inverse logit, exp(x)/(1+exp(x)) = 1/(1+exp(-x))
}
\examples{
expit(2)

}
