% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{print.mmp_prior}
\alias{print.mmp_prior}
\alias{print.mmp_inla}
\alias{print.mmp_stan}
\title{Print}
\usage{
\method{print}{mmp_prior}(x, ...)

\method{print}{mmp_inla}(x, ...)

\method{print}{mmp_stan}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mmp_prior}, \code{mmp_inla} or \code{mmp_stan}.}

\item{...}{For \code{mmp_stan} and \code{mmp_inla}, see \link[base]{print.data.frame}.}
}
\value{
Returns input object invisible.
}
\description{
Print
}
\examples{
\dontrun{
pri <- makemyprior_example_model()
pri # or print(pri)
res_stan <- inference_stan(pri)
res_stan # or print(res_stan)
res_inla <- inference_inla(pri)
res_inla # or print(res_inla)
}

}
