% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_object_functions.R
\name{summary.mmp_prior}
\alias{summary.mmp_prior}
\alias{summary.mmp_inla}
\alias{summary.mmp_stan}
\title{Short summary}
\usage{
\method{summary}{mmp_prior}(object, ...)

\method{summary}{mmp_inla}(object, ...)

\method{summary}{mmp_stan}(object, ...)
}
\arguments{
\item{object}{Object of class \code{mmp_prior}, \code{mmp_inla} or \code{mmp_stan}.}

\item{...}{For \code{mmp_stan} and \code{mmp_inla}, see \link[base]{print.data.frame}.}
}
\value{
Returns summary invisible.
}
\description{
Short summary
}
\examples{
pri <- makemyprior_example_model()
summary(pri)

if (interactive() && requireNamespace("rstan")){
  res_stan <- inference_stan(ex_prior, iter = 100)
  # Note: For reliable results, increase the number of iterations (e.g., 'iter = 2000')
  summary(res_stan)
}

if (interactive() && requireNamespace("INLA")){
  res_inla <- inference_inla(pri)
  summary(res_inla)
}

}
